erwerwerwe

+++ Factoring.Negocios/util/CondicionesDto.cs
@@ -9,50 +9,50 @@ namespace Factoring.Negocios.util
         public int id { get; set; }
 
         /* FACTURING */      
-        public Nullable<float> FactoringPymes { get; set; }
+        public Nullable<decimal> FactoringPymes { get; set; }
 
-        public Nullable<float> FactoringMedianas { get; set; }
+        public Nullable<decimal> FactoringMedianas { get; set; }
 
-        public Nullable<float> FactoringGrandes { get; set; }
+        public Nullable<decimal> FactoringGrandes { get; set; }
 
-        public Nullable<float> FactoringCoorporativas { get; set; }
+        public Nullable<decimal> FactoringCoorporativas { get; set; }
 
-        public Nullable<float> FactoringInmobiliarias { get; set; }
+        public Nullable<decimal> FactoringInmobiliarias { get; set; }
 
         /* CONFIRMING */
-        public Nullable<float> ConfirmingPymes { get; set; }
+        public Nullable<decimal> ConfirmingPymes { get; set; }
 
-        public Nullable<float> ConfirmingMedianas { get; set; }
+        public Nullable<decimal> ConfirmingMedianas { get; set; }
 
-        public Nullable<float> ConfirmingGrandes { get; set; }
+        public Nullable<decimal> ConfirmingGrandes { get; set; }
 
-        public Nullable<float> ConfirmingCoorporativas { get; set; }
+        public Nullable<decimal> ConfirmingCoorporativas { get; set; }
 
-        public Nullable<float> ConfirmingInmobiliarias { get; set; }
+        public Nullable<decimal> ConfirmingInmobiliarias { get; set; }
 
         /* COMBRANZA DELEGADA  */
-        public Nullable<float> CobranzaDelegadaPymes { get; set; }
+        public Nullable<decimal> CobranzaDelegadaPymes { get; set; }
 
-        public Nullable<float> CobranzaDelegadaMedianas { get; set; }
+        public Nullable<decimal> CobranzaDelegadaMedianas { get; set; }
 
-        public Nullable<float> CobranzaDelegadaGrandes { get; set; }
+        public Nullable<decimal> CobranzaDelegadaGrandes { get; set; }
 
-        public Nullable<float> CobranzaDelegadaCoorporativas { get; set; }
+        public Nullable<decimal> CobranzaDelegadaCoorporativas { get; set; }
 
-        public Nullable<float> CobranzaDelegadaInmobiliarias { get; set; }
+        public Nullable<decimal> CobranzaDelegadaInmobiliarias { get; set; }
 
 
-        public Nullable<float> TasaSpreadBolsaProductos { get; set; }
+        public Nullable<decimal> TasaSpreadBolsaProductos { get; set; }
 
-        public Nullable<float> ComisionPyme { get; set; }
+        public Nullable<decimal> ComisionPyme { get; set; }
 
-        public Nullable<float> ComisionBancaEmpresa { get; set; }
+        public Nullable<decimal> ComisionBancaEmpresa { get; set; }
 
         public virtual string ComisionTipoMoneda { get; set; }
 
-        public Nullable<float> GastosNotariaDeudor { get; set; }
+        public Nullable<decimal> GastosNotariaDeudor { get; set; }
 
-        public Nullable<float> GastosNotariaNFacturas { get; set; }
+        public Nullable<decimal> GastosNotariaNFacturas { get; set; }
 
         public virtual string GastosTipoMoneda { get; set; }
